Correcion de logica en servicios y repositorios
