+++ FashionPay.Application/DTOs/Compra/CompraResponseDto.cs
@@ -9,13 +9,7 @@ public class CompraResponseDto
     public int CantidadPagos { get; set; }
     public decimal MontoMensual { get; set; }
     public string? Observaciones { get; set; }
-
-    // Información del cliente
     public ClienteBasicoCompraDto Cliente { get; set; } = new();
-
-    // Productos comprados
     public List<DetalleCompraDto> Detalles { get; set; } = new();
-
-    // Plan de pagos
     public List<PlanPagoDto> PlanPagos { get; set; } = new();
 }
